Change skip_txfm in RD_STATS to a uint8_t
x86 requires specifying the entire immediate value most operations. On a
full x86_64 executable, binary size is reduced by 0.38 KiB.

1) Reorder zero_rate in RD_STATS, so it can be initialized with rate
instead of with skip_txfm (on an instruction level). This is done to
prevent 2) from growing binary size.
2) Change RD_STATS skip_txfm to uint8_t.
3) Change code that depends on skip_txfm's type.
4) Change MB_MODE_INFO's skip_flag from int8_t to uint8_t in an attempt
to unify the type used.

No measurable impact on performance.

+++ av1/encoder/block.h
@@ -496,7 +496,7 @@ typedef struct {
  */
 typedef struct {
   //! Whether to skip transform and quantization on a partition block level.
-  int skip_txfm;
+  uint8_t skip_txfm;
 
   /*! \brief Whether to skip transform and quantization on a txfm block level.
    *
